requirements.txt
black==20.8b1
bokeh==2.2.3
kafka-python==2.0.2
numpy==1.19.5
pandas==1.2.1
pymongo==3.11.2